# finance-manager-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API serves personal finance endpoints. Accuracy is not garunteed",
        "title": "Go Finance Manager API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/users/{userId}/loans": {
            "get": {
                "description": "Returns an array of Loan objects belonging to a given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loans"
                ],
                "summary": "Get All User Loans",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search for loans by name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Loan"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Inserts a new Loan into the Database for a given user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loans"
                ],
                "summary": "Insert Loan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The loan to insert",
                        "name": "loan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Loan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            }
        },
        "/users/{userId}/loans-summary": {
            "get": {
                "description": "Gets a summary of all loans for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loans"
                ],
                "summary": "Get User Loan Summary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LoansSummary"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            }
        },
        "/users/{userId}/loans/{loanId}": {
            "get": {
                "description": "Returns a Loan object belonging to a given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loans"
                ],
                "summary": "Get Loan by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "the ID of the Loan",
                        "name": "loanId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Loan"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing Loan for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loans"
                ],
                "summary": "Update Loan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of the loan to update",
                        "name": "loanId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The loan to update",
                        "name": "loan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Loan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a Loan object belonging to a given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loans"
                ],
                "summary": "Delete Loan by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "the ID of the Loan",
                        "name": "loanId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Loan"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            }
        },
        "/users/{userId}/loans/{loanId}/calculate": {
            "post": {
                "description": "Performs calculations on a loan and returns the loan with updated values.\nDoes not Persist values",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loans"
                ],
                "summary": "Calculate Loan Values",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Loan ID. Will also accept 'new' for unsaved loan",
                        "name": "loanId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The Loan to Calculate values for",
                        "name": "loan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Loan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Loan"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "jsonutils.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Loan": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "interest": {
                    "type": "number"
                },
                "interestRate": {
                    "type": "number"
                },
                "loanTerm": {
                    "type": "integer"
                },
                "monthlyPayment": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "paymentSchedule": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PaymentScheduleItem"
                    }
                },
                "total": {
                    "type": "number"
                },
                "totalCost": {
                    "type": "number"
                },
                "totalPayment": {
                    "type": "number"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "models.LoansSummary": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "monthlyCost": {
                    "type": "number"
                },
                "totalBalance": {
                    "type": "number"
                }
            }
        },
...
        "models.PaymentScheduleItem": {
            "type": "object",
            "properties": {
                "interest": {
                    "type": "number"
                },
                "interestToDate": {
                    "type": "number"
                },
                "month": {
                    "type": "integer"
                },
                "principal": {
                    "type": "number"
                },
                "principalToDate": {
                    "type": "number"
                },
                "remainingBalance": {
                    "type": "number"
                }
            }
        },
...
    }
}